Create Ebook
I have created an ebook that will help freshers and job seekers to build a strong resume and stand out in shortlisting. 
In this ebook, I have also shared a special trick to get interview calls from LinkedIn and tips on how to behave during interviews with confidence. 
It is written in simple language so anyone can understand and apply these steps easily.